Refactor code: split Help to ReportService and StudioService, implement 2 interfaces for each service Fix: When including/removing single file projects, check the languages only when at least one project is selected
