+++ Export Analysis Reports/Sdl.Community.ExportAnalysisReports/Sdl.Community.ExportAnalysisReports.csproj
@@ -120,6 +120,8 @@
     <Compile Include="Helpers\Reflector.cs" />
     <Compile Include="Helpers\StudioAnalysisReport.cs" />
     <Compile Include="Interfaces\IMessageBoxService.cs" />
+    <Compile Include="Interfaces\IReportService.cs" />
+    <Compile Include="Interfaces\IStudioService.cs" />
     <Compile Include="Model\AnalyzedFile.cs" />
     <Compile Include="Model\BandResult.cs" />
     <Compile Include="Model\JsonSettings.cs" />
@@ -141,6 +143,8 @@
     </Compile>
     <Compile Include="ReportExporterRibbon.cs" />
     <Compile Include="Service\MessageBoxService.cs" />
+    <Compile Include="Service\ReportService.cs" />
+    <Compile Include="Service\StudioService.cs" />
     <Compile Include="TellMe\CommunitySupportAction.cs" />
     <Compile Include="TellMe\HelpAction.cs" />
     <Compile Include="TellMe\StoreAction.cs" />
